refactor / add to genlims & dbify tests; need to figure out how best to scope mock_db in confest
