+++ bripipetools/genlims/__init__.py
@@ -2,7 +2,8 @@
 Interactions with the GenLIMS Mongo database.
 """
 from .connection import db
-from .operations import (get_samples, get_runs, get_workflowbatches,
+from .operations import (find_objects, insert_objects,
+                         get_samples, get_runs, get_workflowbatches,
                          put_samples, put_runs, put_workflowbatches,
                          create_workflowbatch_id)
 from .mapping import (map_keys, get_class, map_to_object)
