+++ tests/conftest.py
@@ -4,6 +4,7 @@ logger = logging.getLogger(__name__)
 import os
 
 import pytest
+import mongomock
 
 @pytest.fixture(scope='class')
 def mock_genomics_server(request):
@@ -71,3 +72,46 @@ def mock_genomics_server(request):
                      "from mock 'genomics' server"))
     request.addfinalizer(fin)
     return data
+
+@pytest.fixture(scope='class')
+def mock_db(request):
+    logger.info(("[setup] mock 'tg3' database, connect "
+                 "to mock 'tg3' Mongo database"))
+    db = mongomock.MongoClient().db
+    db.samples.insert(
+        {"_id": "lib7293",
+    	"projectId": 14,
+    	"projectName": "U01-Mexico 2011",
+    	"sampleId": "S2733",
+    	"libraryId": "lib7293",
+    	"parentId": "grRNA5942",
+    	"type": "library"}
+    )
+    db.runs.insert(
+        {"_id": "150615_D00565_0087_AC6VG0ANXX",
+        "date": "2015-06-15",
+    	"instrumentId": "D00565",
+    	"runNumber": 87,
+    	"flowcellId": "C6VG0ANXX",
+    	"flowcellPosition": "A",
+    	"type": "flowcell"}
+    )
+    db.workflowbatches.insert(
+        {"_id": "globusgalaxy_2016-04-12_1",
+        "workflowbatchFile": ("/genomics/Illumina/"
+                              "150615_D00565_0087_AC6VG0ANXX/"
+                              "globus_batch_submission/"
+                              "160412_P109-1_P14-12_C6VG0ANXX_"
+                              "optimized_truseq_unstrand_sr_grch38_"
+                              "v0.1_complete.txt"),
+        "date": "2016-04-12",
+    	"workflowId": "optimized_truseq_unstrand_sr_grch38_v0.1_complete.txt",
+    	"projects": ["P109-1", "P14-12"],
+    	"flowcellId": "C6VG0ANXX",
+        "type": "Galaxy workflow batch"}
+    )
+    def fin():
+        logger.info(("[teardown] mock 'tg3' database, disconnect "
+                     "from mock 'tg3' Mongo database"))
+    request.addfinalizer(fin)
+    return db
