+++ tests/test_dbify.py
@@ -11,38 +11,6 @@ from bripipetools import model as docs
 from bripipetools import dbify
 
 @pytest.fixture(scope='class')
-def mock_genomics_server(request):
-    logger.info(("[setup] mock 'genomics' server, connect "
-                 "to mock 'genomics' server"))
-    run_id = '150615_D00565_0087_AC6VG0ANXX'
-    mock_genomics_root = './tests/test-data/'
-    mock_genomics_path = os.path.join(mock_genomics_root, 'genomics')
-    mock_flowcell_path = os.path.join(mock_genomics_path, 'Illumina', run_id)
-    mock_unaligned_path = os.path.join(mock_flowcell_path, 'Unaligned')
-    mock_workflow_batch_file = os.path.join(
-        mock_flowcell_path, 'globus_batch_submission',
-        ("160412_P109-1_P14-12_C6VG0ANXX_"
-         "optimized_truseq_unstrand_sr_grch38_v0.1_complete.txt")
-    )
-
-    data = {'run_id': run_id,
-            'genomics_root': mock_genomics_root,
-            'genomics_path': mock_genomics_path,
-            'flowcell_path': mock_flowcell_path,
-            'unaligned_path': mock_unaligned_path,
-            'project_p14_12': 'P14-12-23221204',
-            'project_p109_1': 'P109-1-21113094',
-            'lib7293': 'lib7293-25920016',
-            'lib7293_fastq': 'MXU01-CO072_S1_L001_R1_001.fastq.gz',
-            'workflowbatch_file': mock_workflow_batch_file,
-            'batch_name': '160412_P109-1_P14-12_C6VG0ANXX'}
-    def fin():
-        logger.info(("[teardown] mock 'genomics' server, disconnect "
-                     "from mock 'genomics' server"))
-    request.addfinalizer(fin)
-    return data
-
-@pytest.fixture(scope='class')
 def mock_db(request):
     logger.info(("[setup] mock 'tg3' database, connect "
                  "to mock 'tg3' Mongo database"))
@@ -173,7 +141,7 @@ class TestSequencingImporter:
         # WHEN inserting with collection argument set to 'samples'
         importer.insert(collection='samples')
 
-        # THEN documents should be present in only runs collections
+        # THEN new documents should be present in only runs collections
         assert(len(list(mock_db.runs.find({'type': 'flowcell'})))
                == 0)
         assert(len(list(mock_db.samples.find({'type': 'sequenced library'})))
